Main Gameplay Loop Update
Should be able to scan items, take payment, and give items and receipt to finish transaction.
Known bugs:
Receipt does not reset upon being given, so can not be given again.
Sometimes, when items are supposed to spawn there is a crash.
NullPointerException when all money is taken, does not cause any crash or change in gameplay; safe to ignore.
Correct amount of money always spawns, but sometimes in odd amounts, like 5 pennies instead of a nickel.